re #262.
- In sentences_controller.php: added a return in case no id is provided for "delete".
- Added delete.ctp.
- In helper/lists.php: don't display the sentence in the list if it's a deleted sentence. This will also prevent moderators from trying to delete sentences that have already been deleted.

+++ app/controllers/sentences_controller.php
@@ -254,9 +254,13 @@ class SentencesController extends AppController
      *
      * @return void
      */
-    public function delete($id)
+    public function delete($id = null)
     {
         $id = Sanitize::paranoid($id);
+        if (empty($id)) {
+            return;
+        }
+        
         $this->Sentence->delete(
             $id,
             true
